# openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Agent Platform API
  version: 1.0.0
  description: |
    REST API contract for the Multi-Agent Platform backend.

    This document is the single source of truth shared between backend (Spring Boot 4.0.6 / Java 17 / Spring AI 1.1.0)
    and frontend (ReactJS). It is derived from `backend/requirements/REQS.md` and `backend/design/SW-DESIGN.md`.

    ## Authentication
    Two mutually exclusive modes are supported:
    - **JWT mode** (end users) — `Authorization: Bearer <token>` issued by `POST /auth/login`.
      Default lifetime: 30 minutes (HS256). No refresh tokens. Logout adds the token's `jti` to a
      self-expiring denylist.
    - **API-key mode** (machine-to-machine) — `X-Client-Id` and `X-Api-Key` headers.
      API-key callers run under a virtual `SYSTEM` principal with full chat capabilities but no
      admin access and no visibility on end-user-owned resources.

    ## Errors
    All error responses use RFC 7807 Problem Details with a stable machine-readable `code` field.

    ## Pagination
    List endpoints use opaque cursor-based pagination (`cursor`, `pageSize`) with a `nextCursor`
    in the response envelope. Default page size is 20, maximum 100.

    ## Streaming
    The chat endpoint `POST /conversations/{conversationId}/messages` returns Server-Sent Events
    (`text/event-stream`) with typed frames: `started`, `delta`, `completed`, `error`.
  contact:
    name: Multi-Agent Platform Team
...
  license:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Sign in with email and password
      description: |
        Public endpoint. Returns a signed JWT to be used in the `Authorization: Bearer <token>` header
        on subsequent calls. The response includes `mustChangePassword` so the frontend can route the
        user to the password-change screen when required (REQ-USR-007).
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/InvalidCredentials'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log the current JWT out (denylist)
      description: |
        Adds the presented JWT's `jti` to the logout denylist. The denylist entry self-expires no later
        than the token's natural `exp`. Subsequent requests using the same token are rejected with 401.
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Logged out. The client must also discard the JWT locally.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/password:
    put:
      tags: [Auth]
      summary: Change the authenticated user's password
      description: |
        Authenticated end-user endpoint. The new password must meet the platform policy
        (≥10 chars, ≥1 uppercase, ≥1 special char). On success, the `mustChangePassword` flag
        is cleared. The current JWT remains valid until natural expiry.
      operationId: changeOwnPassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

  # ------------------------------------------------------------
  # Admin - Users
  # ------------------------------------------------------------
...
  /admin/rate-limit:
    get:
      tags: [Admin - Rate Limit]
      summary: Get the live rate-limit configuration
      operationId: getRateLimitConfig
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current rate-limit configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Admin - Rate Limit]
      summary: Replace the live rate-limit configuration
      description: Takes effect on the next request without restart.
      operationId: updateRateLimitConfig
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RateLimitConfigRequest'
      responses:
        '200':
          description: Updated rate-limit configuration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

  # ------------------------------------------------------------
  # Catalogues - Tools / MCP Servers
  # ------------------------------------------------------------
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT issued by `POST /auth/login`. HS256-signed. Default lifetime 30 minutes.
        Submitted as `Authorization: Bearer <token>`.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Machine-to-machine credentials issued by an admin via `POST /admin/api-keys`.
        Both `X-Api-Key` and `X-Client-Id` headers must be present on every request.
        OpenAPI 3.0 lacks composite api-key schemes; consumers SHALL also send `X-Client-Id`.

  # ----------------------------------------------------------
  # Reusable parameters
  # ----------------------------------------------------------
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            validation:
              value:
                type: https://errors.multi-agent-platform/validation-error
                title: Validation error
                status: 400
                detail: One or more fields failed validation.
                code: VALIDATION_ERROR
                errors:
                  - field: name
                    message: must be at most 32 characters
    Unauthorized:
      description: Authentication missing, invalid, expired, or denylisted.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            invalidCredentials:
              value:
                type: https://errors.multi-agent-platform/invalid-credentials
                title: Invalid credentials
                status: 401
                detail: Authentication failed.
                code: INVALID_CREDENTIALS
    InvalidCredentials:
      description: Generic authentication failure (no leak about email existence or credential format).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            invalidCredentials:
              value:
                type: https://errors.multi-agent-platform/invalid-credentials
                title: Invalid credentials
                status: 401
                detail: Authentication failed.
                code: INVALID_CREDENTIALS
    Forbidden:
      description: |
        The authenticated principal is not allowed to perform this action.
        Also returned with code `MUST_CHANGE_PASSWORD` when the user must change their password
        before any other operation is permitted.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            forbidden:
              value:
                type: https://errors.multi-agent-platform/forbidden
                title: Forbidden
                status: 403
                detail: You are not allowed to perform this action.
                code: FORBIDDEN
            mustChangePassword:
              value:
                type: https://errors.multi-agent-platform/must-change-password
                title: Password change required
                status: 403
                detail: Change your password before performing any other operation.
                code: MUST_CHANGE_PASSWORD
...
    RateLimited:
      description: Global rate limit exceeded.
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying.
          schema:
            type: integer
            minimum: 0
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            rateLimited:
              value:
                type: https://errors.multi-agent-platform/rate-limited
                title: Too many requests
                status: 429
                detail: Global rate limit exceeded; retry later.
                code: RATE_LIMITED
    InternalError:
      description: Unexpected server error.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          examples:
            internal:
              value:
                type: https://errors.multi-agent-platform/internal-error
                title: Internal error
                status: 500
                detail: An unexpected error occurred.
                code: INTERNAL_ERROR

  # ----------------------------------------------------------
  # Schemas
  # ----------------------------------------------------------
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807 Problem Details with stable machine-readable `code`.
      required: [status, title, code]
      properties:
        type:
          type: string
          format: uri
          description: URI identifier of the error type.
        title:
          type: string
          description: Short, human-readable summary.
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: HTTP status code.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: Path of the request that triggered the error.
        code:
          type: string
          description: Stable machine-readable error code.
          enum:
            - VALIDATION_ERROR
            - INVALID_CREDENTIALS
            - MUST_CHANGE_PASSWORD
            - FORBIDDEN
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
            - CONFLICT
            - DUPLICATE_AGENT_NAME
            - NESTED_TEAM_FORBIDDEN
            - CROSS_OWNER_TEAM_MEMBER
            - CONVERSATION_FULL
            - RATE_LIMITED
            - LLM_UNAVAILABLE
            - MCP_SERVER_ERROR
            - NOT_ACCEPTABLE
            - INTERNAL_ERROR
        errors:
          type: array
          description: Optional per-field validation details.
          items:
            type: object
            required: [field, message]
            properties:
              field:
                type: string
              message:
                type: string
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 256
          writeOnly: true
    LoginResponse:
      type: object
      required: [token, tokenType, expiresAt, mustChangePassword]
      properties:
        token:
          type: string
          description: Signed JWT (HS256).
        tokenType:
          type: string
          enum: [Bearer]
        expiresAt:
          type: string
          format: date-time
          description: Token expiry, ISO 8601 UTC.
        mustChangePassword:
          type: boolean
          description: True if the user must change their password before any other operation.
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          format: password
          writeOnly: true
        newPassword:
          type: string
          format: password
          minLength: 10
          maxLength: 256
          writeOnly: true
          description: |
            Must satisfy the platform password policy:
            ≥10 characters, ≥1 uppercase letter, ≥1 special character.

    # ----- Users ----------------------------------------------------------
...
    RateLimitConfig:
      type: object
      required: [perMinute, perHour, updatedAt]
      properties:
        perMinute:
          type: integer
          minimum: 1
          description: Maximum requests per minute (global bucket).
        perHour:
          type: integer
          minimum: 1
          description: Maximum requests per hour (global bucket).
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
          nullable: true
          description: Admin user identifier who last updated the configuration.
    RateLimitConfigRequest:
      type: object
      required: [perMinute, perHour]
      properties:
        perMinute:
          type: integer
          minimum: 1
        perHour:
          type: integer
          minimum: 1

    # ----- Tools / MCP catalog -------------------------------------------
...